backend/requirements.txt
fastapi
pymongo
uvicorn[standard]
paho-mqtt
redis
fastapi-login
python-jose[cryptography]
python-dotenv
passlib[bcrypt]
PyJWT
strawberry-graphql[fastapi]
redis
influxdb
dacite
